+++ openbank-release-steward/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [0.6.0](https://github.com/JiRaska/open-bank-oss/compare/release-steward-v0.5.0...release-steward-v0.6.0) (2026-08-06)
+
+
+### Features
+
+* **release-steward:** wire prompt registry and LLM gateway ([#3785](https://github.com/JiRaska/open-bank-oss/issues/3785)) ([17679d7](https://github.com/JiRaska/open-bank-oss/commit/17679d7b6328208c6845e41a13ab6e034a4ce680)), closes [#3784](https://github.com/JiRaska/open-bank-oss/issues/3784)
+
+## [0.5.0](https://github.com/JiRaska/open-bank-oss/compare/release-steward-v0.4.1...release-steward-v0.5.0) (2026-08-05)
+
+
+### Features
+
+* **scheduler:** register workflow liveness for second non-money-path batch (ADR-0237) ([#3739](https://github.com/JiRaska/open-bank-oss/issues/3739)) ([735e8bd](https://github.com/JiRaska/open-bank-oss/commit/735e8bdc12fbf541464aeb4f15ce767cb7866e78))
+
 ## [0.4.1](https://github.com/JiRaska/open-bank-oss/compare/release-steward-v0.4.0...release-steward-v0.4.1) (2026-08-03)